Implemented a bootstrap class following Johns instruction, which appears to have introduced me to the concept of Application Listeners and Events in general. I feel like I have a lot to learn here haha.
Anyway, the crud repos we added allowed us to save some data to the H2 database super easy, which I'm impressed with.
I know that the importance of the Repository layer is that it potentially lets us swap thing out in future, in terms of the DBs, so its pretty important to understand this layer well.